// src/models/chatmessages.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 * @generated-id: f21e02796127
 */

import * as z from "zod/v4";
import {
  ChatAssistantMessage,
  ChatAssistantMessage$Outbound,
  ChatAssistantMessage$outboundSchema,
} from "./chatassistantmessage.js";
import {
  ChatDeveloperMessage,
  ChatDeveloperMessage$Outbound,
  ChatDeveloperMessage$outboundSchema,
} from "./chatdevelopermessage.js";
import {
  ChatSystemMessage,
  ChatSystemMessage$Outbound,
  ChatSystemMessage$outboundSchema,
} from "./chatsystemmessage.js";
import {
  ChatToolMessage,
  ChatToolMessage$Outbound,
  ChatToolMessage$outboundSchema,
} from "./chattoolmessage.js";
import {
  ChatUserMessage,
  ChatUserMessage$Outbound,
  ChatUserMessage$outboundSchema,
} from "./chatusermessage.js";

/**
 * Chat completion message with role-based discrimination
 */
export type ChatMessages =
  | (ChatAssistantMessage & { role: "assistant" })
  | ChatDeveloperMessage
  | ChatSystemMessage
  | ChatToolMessage
  | ChatUserMessage;

/** @internal */
export type ChatMessages$Outbound =
  | (ChatAssistantMessage$Outbound & { role: "assistant" })
  | ChatDeveloperMessage$Outbound
  | ChatSystemMessage$Outbound
  | ChatToolMessage$Outbound
  | ChatUserMessage$Outbound;

/** @internal */
export const ChatMessages$outboundSchema: z.ZodType<
  ChatMessages$Outbound,
  ChatMessages
> = z.union([
  ChatAssistantMessage$outboundSchema.and(
    z.object({ role: z.literal("assistant") }),
  ),
  ChatDeveloperMessage$outboundSchema,
  ChatSystemMessage$outboundSchema,
  ChatToolMessage$outboundSchema,
  ChatUserMessage$outboundSchema,
]);

export function chatMessagesToJSON(chatMessages: ChatMessages): string {
  return JSON.stringify(ChatMessages$outboundSchema.parse(chatMessages));
}