// src/models/chattoolmessage.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v4";
import { remap as remap$ } from "../lib/primitives.js";
import {
  ChatContentItems,
  ChatContentItems$Outbound,
  ChatContentItems$outboundSchema,
} from "./chatcontentitems.js";

/**
 * Tool response content
 */
export type ChatToolMessageContent = string | Array<ChatContentItems>;

/**
 * Tool response message
 */
export type ChatToolMessage = {
  role: "tool";
  /**
   * Tool response content
   */
  content: string | Array<ChatContentItems>;
  /**
   * ID of the assistant message tool call this message responds to
   */
  toolCallId: string;
};

/** @internal */
export type ChatToolMessageContent$Outbound =
  | string
  | Array<ChatContentItems$Outbound>;

/** @internal */
export const ChatToolMessageContent$outboundSchema: z.ZodType<
  ChatToolMessageContent$Outbound,
  ChatToolMessageContent
> = z.union([z.string(), z.array(ChatContentItems$outboundSchema)]);

export function chatToolMessageContentToJSON(
  chatToolMessageContent: ChatToolMessageContent,
): string {
  return JSON.stringify(
    ChatToolMessageContent$outboundSchema.parse(chatToolMessageContent),
  );
}

/** @internal */
export type ChatToolMessage$Outbound = {
  role: "tool";
  content: string | Array<ChatContentItems$Outbound>;
  tool_call_id: string;
};

/** @internal */
export const ChatToolMessage$outboundSchema: z.ZodType<
  ChatToolMessage$Outbound,
  ChatToolMessage
> = z.object({
  role: z.literal("tool"),
  content: z.union([z.string(), z.array(ChatContentItems$outboundSchema)]),
  toolCallId: z.string(),
}).transform((v) => {
  return remap$(v, {
    toolCallId: "tool_call_id",
  });
});

export function chatToolMessageToJSON(
  chatToolMessage: ChatToolMessage,
): string {
  return JSON.stringify(ChatToolMessage$outboundSchema.parse(chatToolMessage));
}